Image Rebase + Image Enhancement
Added image enhancement which subsequently saves more images. To avoid cluttering the root folder I've added an image manipulation folder to handle the saves.
Enhancing the images produces a clearer and crisper image which the model seems to get a bit easier
It still really struggles on 8, it has yet to produce one.
0-7 are quite accurate. 9 is so so.