Refactor app.py
Completely refactored app.py so that it is better layed out and is broken down into re-usable methods.
Only thing that isn't re-usable is the design of the individual image enhance functions, each one depends on the other with hardcoded variables names. Need to change that.
Commented and explained what each method does, along with a reference to where and how I figured it out